Reorganize files, create headers, format headers
