+++ src/spawner/modules/randommap_hooks.cpp
@@ -1,22 +1,33 @@
-///**
-//*  yrpp-spawner
-//*
-//*  Copyright(C) 2022-present CnCNet
-//*
-//*  This program is free software: you can redistribute it and/or modify
-//*  it under the terms of the GNU General Public License as published by
-//*  the Free Software Foundation, either version 3 of the License, or
-//*  (at your option) any later version.
-//*
-//*  This program is distributed in the hope that it will be useful,
-//*  but WITHOUT ANY WARRANTY; without even the implied warranty of
-//*  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.See the
-//*  GNU General Public License for more details.
-//*
-//*  You should have received a copy of the GNU General Public License
-//*  along with this program.If not, see <http://www.gnu.org/licenses/>.
-//*/
-//
+/*******************************************************************************
+/*                 O P E N  S O U R C E  --  V I N I F E R A                  **
+/*******************************************************************************
+ *
+ *  @project       Vinifera
+ *
+ *  @file          RANDOMMAP_HOOKS.CPP
+ *
+ *  @author        ZivDero
+ *
+ *  @brief         Contains the hooks for the random map generator.
+ *
+ *  @license       Vinifera is free software: you can redistribute it and/or
+ *                 modify it under the terms of the GNU General Public License
+ *                 as published by the Free Software Foundation, either version
+ *                 3 of the License, or (at your option) any later version.
+ *
+ *                 Vinifera is distributed in the hope that it will be
+ *                 useful, but WITHOUT ANY WARRANTY; without even the implied
+ *                 warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
+ *                 PURPOSE. See the GNU General Public License for more details.
+ *
+ *                 You should have received a copy of the GNU General Public
+ *                 License along with this program.
+ *                 If not, see <http://www.gnu.org/licenses/>.
+ *
+ ******************************************************************************/
+
+#include "randommap_hooks.h"
+
 //#include "Spawner.h"
 //
 //#include <GameStrings.h>
