+++ src/spawner/net/nethack.cpp
@@ -1,26 +1,46 @@
-/*
-* Copyright (c) 2012, 2013, 2014 Toni Spets <toni.spets@iki.fi>
-*
-* Permission to use, copy, modify, and distribute this software for any
-* purpose with or without fee is hereby granted, provided that the above
-* copyright notice and this permission notice appear in all copies.
-*
-* THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
-* WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
-* MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
-* ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
-* WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
-* ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
-* OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
-*/
+/*******************************************************************************
+/*                 O P E N  S O U R C E  --  V I N I F E R A                  **
+/*******************************************************************************
+ *
+ *  @project       Vinifera
+ *
+ *  @file          NETHACK.CPP
+ *
+ *  @author        Toni Spets
+ *
+ *  @brief         Nethack.
+ *
+ *  @license       Vinifera is free software: you can redistribute it and/or
+ *                 modify it under the terms of the GNU General Public License
+ *                 as published by the Free Software Foundation, either version
+ *                 3 of the License, or (at your option) any later version.
+ *
+ *                 Vinifera is distributed in the hope that it will be
+ *                 useful, but WITHOUT ANY WARRANTY; without even the implied
+ *                 warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
+ *                 PURPOSE. See the GNU General Public License for more details.
+ *
+ *                 You should have received a copy of the GNU General Public
+ *                 License along with this program.
+ *                 If not, see <http://www.gnu.org/licenses/>.
+ *
+ *  @note          Copyright (c) 2012, 2013, 2014 Toni Spets <toni.spets@iki.fi>
+ *
+ *                 Permission to use, copy, modify, and distribute this
+ *                 software for any purpose with or without fee is hereby
+ *                 granted, provided that the above copyright notice and
+ *                 this permission notice appear in all copies.
+ *
+ ******************************************************************************/
 
-#include "NetHack.h"
-#include "Spawner.h"
+#include "nethack.h"
 
 #include <windows.h>
-#include <stdint.h>
+#include <cstdint>
 #include <winsock2.h>
 #pragma comment(lib, "wsock32.lib")
+
+#include "spawner.h"
 
 ListAddress ListAddress::Array[8] = {};
